# openapi.yaml
openapi: 3.0.0
info:
  title: sp_internal_api
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /docs:
    get:
      tags:
        - Docs
      summary: Get the OpenAPI specification
      responses:
        "200":
          description: OpenAPI specification
  /health:
    get:
      tags:
        - Health
      summary: Check the health of the server
      responses:
        "200":
          description: Server is up and running
  /config:
    get:
      tags:
        - Config
      summary: Get the configuration of webapp
      responses:
        "200":
          description: Configuration of webapp
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GetConfigResponse"
  /tables:
    post:
      summary: Create a new table
      tags:
        - Tables
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableDto"
      responses:
        "200":
          description: Table created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TableEntity"
        "400":
          description: Table already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: Get table by parameters
      tags:
        - Tables
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
          description: ID of the user
        - name: type_id
          in: query
          required: true
          schema:
            type: integer
          description: ID of the table type
        - name: month_year
          in: query
          required: true
          schema:
            type: string
            example: "2025-02"
          description: Month and year in YYYY-MM format
      responses:
        "200":
          description: Table retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GetTableResponse"
        "404":
          description: Table not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tables/{id}:
    get:
      tags:
        - Tables
      summary: Get table by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GetTableResponse"
        "404":
          description: Table not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions:
    post:
      summary: Create a new transaction
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionDto"
      responses:
        "200":
          description: Transaction created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TransactionEntity"
        "404":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    CreateTableDto:
      type: object
      properties:
        user_id:
          type: integer
        type_id:
          type: integer
        month_year:
          type: string
      required:
        - user_id
        - type_id
        - month_year
      example:
        user_id: 1
        type_id: 1
        month_year: "2023-01"
    CreateTransactionDto:
      type: object
      properties:
        description:
          type: string
        table_id:
          type: integer
        type_id:
          type: integer
        amount:
          type: number
        currency_id:
          type: integer
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      example:
        message: Error message
        error: Error details
        success: false
    CurrencyEntity:
      type: object
      properties:
        id:
          type: integer
        description: 
          type: string
        abbreviation:
          type: string
        symbol:
          type: string
        status:
          type: boolean 
        created_at:
          type: string
        updated_at:
          type: string
    TableEntity:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        type_id:
          type: integer
        amount:
          type: number
        status:
          type: boolean
        month_year:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    TransactionEntity:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        type_id:
          type: integer
        amount:
          type: number
        currency_id:
          type: integer
        status:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
    TransactionTypeEntity:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        description:
          type: string
        status:
          type: boolean
        table_type_id:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
    TableTypeEntity:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        description:
          type: string
        status:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
    GetTableResponse:
      type: object
      properties:
        table:
          $ref: "#/components/schemas/TableEntity"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionEntity"
    GetConfigDatesResponse:
      type: object
      properties:
        months:
          type: array
          items:
            type: integer
        years:
          type: array
          items:
            type: integer
    GetConfigResponse:
      type: object
      properties:
        currencies:
          type: array
          items:
            $ref: "#/components/schemas/CurrencyEntity"
        transaction_types:
          type: array
          items:
            $ref: "#/components/schemas/TransactionTypeEntity"
        month_years:
          $ref: "#/components/schemas/GetConfigDatesResponse"
        table_types:
          type: array
          items:
            $ref: "#/components/schemas/TableTypeEntity"